Check if nifti image or read in a nifti image

Simple check to see if input is character or of class nifti

check_nifti(x, reorient = FALSE, allow.array = FALSE, fast = FALSE,
  need_header = TRUE)

# S4 method for nifti
check_nifti(x, reorient = FALSE, allow.array = FALSE,
  fast = FALSE, need_header = TRUE)

# S4 method for character
check_nifti(x, reorient = FALSE, allow.array = FALSE,
  fast = FALSE, need_header = TRUE)

# S4 method for factor
check_nifti(x, reorient = FALSE, allow.array = FALSE,
  fast = FALSE, need_header = TRUE)

# S4 method for list
check_nifti(x, reorient = FALSE, allow.array = FALSE,
  fast = FALSE, need_header = TRUE)

# S4 method for array
check_nifti(x, reorient = FALSE, allow.array = FALSE,
  fast = FALSE, need_header = TRUE)

# S4 method for anlz
check_nifti(x, reorient = FALSE, allow.array = FALSE,
  fast = FALSE, need_header = TRUE)

# S4 method for ANY
check_nifti(x, reorient = FALSE, allow.array = FALSE,
  fast = FALSE, need_header = TRUE)

Arguments

x

character path of image or an object of class nifti, or array

reorient

(logical) passed to readnii if the image is to be re-oriented

allow.array

(logical) Are array types allowed (TRUE) or should there be an error if the object is not character or class nifti.

fast

if TRUE, then fast_readnii will be used versus readnii if the files need to be read in.

need_header

if TRUE, then an image type with header information will be returned. If not, then an array is fine. Used really only in conjunction with allow.array

Value

nifti object or array if allow.array=TRUE and x is an array

See also

readnii