Double Orthographic Display

Copy of oro.nifti's orthographic function with some tweaks such as adding L/R designations for left and right

double_ortho(x, y = NULL, col.y = gray(0:64/64), NA.x = TRUE,
  mfrow = c(2, 4), add = FALSE, ...)

Arguments

x

is an object of class nifti or similar.

y

is an object of class nifti or similar to be set aside x.

col.y

is grayscale (by default).

NA.x

Set any values of 0 in x to NA

mfrow

(numeric) layout of the 3 slices

add

Should the y-plot be added or its own plot? Used in double_ortho

...

other arguments to ortho2

See also

orthographic