Image Time Series to list

Turns a 4D time series image to a list of 3D images

img_ts_to_list(imgs, copy_nifti = TRUE, warn = TRUE)

Arguments

imgs

object of class nifti with 4 dimensions, aka a 4D time series

copy_nifti

Should nifti objects be returned (TRUE) or simply arrays (FALSE). Should only be used for slight speed up when array is adequate

warn

Should a warning be printed if object is not class nifti

Value

List of images

Note

If the object is not of class nifti or have 4 dimensions, then the object is returned