Window image based on quantiles of Image

Takes an image, finds the quantiles given by probs, then sets values outside these values to other values, as determined by replace argument passed to to window_img

robust_window(img, non_zero = FALSE, probs = c(0, 0.999), ..., mask = NULL)

Arguments

img
object of class nifti

non_zero
Should zeroes be excluded from the calculation of quantiles?

probs
quantiles to constrain the image these define the window sent to window_img

...
additional arguments sent to window_img

mask
binary image to use to to calculate quantiles

Value

Object of class nifti with values outside quantiles replaced by values depending on replace argument passed to window_img